runner/pyproject.toml
[tool.poetry]
name = "runner"
version = "0.1.0"
description = "Good-Bot's command runner."
authors = ["TrickyTroll <tricky@beon.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
pexpect = "^4.8.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
runner = "runner.runner:main"